api: improve bolthold usage
We appreciate your understanding of the large diff.

Bolthold allows us to do bolt transactions, which help us maintain
data integrity. We have rewritten several database operations
using transactions.

Bolthold is a low-level tool not suitable for direct use in route
handlers, which need a distinct vocabulary of operations to support
their job. We need to flesh out the operations involving it before
we can pack them up nicely. That said, we want to keep all the dirty
work inside api/db.go and out of files that serve HTTP requests,
such as api/device.go and api/user.go.

The api.Db type is nearly superfluous at this point, given that many
of its methods can be converted into functions of bolthold.Store. We
intend to remove it completely soon. Something like it may return
in the future.

As we learn to make good use of bolthold, we make two observations:

1. There's more code. There may be opportunities to reduce it. We
will carefully consider this.

2. Data embedding can be useful.

Differences bewteen wire and storage types, as entailed by the
current design (mostly relational schema, server-side joins), have
raised subtle questions. For example, to update an existing user,
we must update the set of roles. One approach is to analyze the
difference between old and new role sets, which seems difficult.

Another approach is to delete all the old roles and create new ones,
with new indexes. This assumes that the roles are never referred to
elsewhere, which is currently true, but may not be in the future,
as the schema makes it possible.

An embedded schema can ease the task of updating this data. By
embedding roles, we can codify that they cannot be "referred" to
by anything else. This also reduces the number of joins required
to serve GET requests.

We expect the risk of this embedding to be low, as we can't think
of a use for referring to roles.

Signed-off-by: Karl Schultheisz <kdsch@protonmail.com>