globalize frontend state; distinguish root user/org
Backend changes:

The backend responds to a login request with JSON containing the auth
token (as previously) and the user's privilege level. A failed login
has no privilege. The other possibilities are user, admin, and root.
The frontend validates but does not currently use privilege
information.

We enforce the policy that all devices are properties of root upon
a new device posting a sample.

The root org and user are statically assigned a zero UUID to aid
distinguishing them from other entities.

We use a transaction to check initialization of the datastore.

We added a /v1/data endpoint, which serves users, orgs, and devices
in one shot.

Frontend changes:

Device, org, and user lists have been moved to Global.Model.

Upon login, we update the global model by requesting /v1/data.

The device list update ticker has been moved to Global. The ticker
updates the device list constantly.

We've moved code for user, org, and device types into their own
modules:

	                  Global
	                 /  |  \
	                v   v   v
	           Device  User  Org
	           ^        ^      ^
	          /         |       \
	Pages.Devices Pages.Users Pages.Orgs

Page-local requests to update the org and user lists still need to be
removed.

Signed-off-by: Karl Schultheisz <kdsch@protonmail.com>